feat(#83): Modularize frontend components
The image-upload component is now renamed to book-search-component, and
contains three components which perform the actual search and upload
functionality:
- image-search-component displays the picture and sends the search
request to the backend
- booklist-component displays the found results and provides
functionality for selecting certain books. It has as an input field a
class that implements the AddTo interface that it will call when the
selected books are submitted.
- manual-book-find-component provides functionality for manually adding
a book to the users collection

The image-search and booklist-component are now also reused in the
recommendation-component