bug(#88): Fix Tesseract smoke test
This change fixes a test in the `BasicTesseractExampleTest` class. This
test failed due to Leptonica failing to read a .png image directly from
a byte array. Leptonica would try to write this buffer into a temporary file,
but failed to implement this correctly on Windows systems. As
workaround, we write the buffer ourselves to a temporary file and let
Leptonica read from the file instead.