chore(#80): Ignore Kotlin-generated code in coverage report
This change will make Jacoco ignore Kotlin-generated code in the
coverage report. This feature was added recently in Jacoco and requires
us to use a snapshot version of Jacoco.

The benefit of this approach is that this does not require any compiler
plugins and which implies the build will keep on working in IntelliJ as
opposed to !47.