refactor(#106): Make API interfaces more consistent
This change makes the API interfaces by the backend server more
consistent. This means the (error) responses are unified  across the
endpoints (except for the OAuth endpoint for compliance reasons).

This change defines the `Success<T>` and `Failure` classes which are
used to represent an API response from the backend server.