refactor(#107): Remove static Kotlin Spring configuration
This change removes the static Spring configuration in Kotlin. Namely, the
`BooklabConfiguration` class. This class allows the user to configure the
Spring container during compile time in addition
to the XML configuration.

We want to remove this feature since it is not used a lot (almost all
services are already configured in XML) and adds confusion due to there
being multiple places where the user can configure the Spring beans.
This issue aims to create one unified place for configuring the beans.

For existing beans in the Kotlin configuration that cannot be translated
to a XML definition and need to be constructed manually, we can write
`FactoryBean<T>` instances to assist with the construction.