refactor(#107): Add alternative deployment types
This change adds two additional deployment types to the testing
deployment: production and oss.

The production deployment type is a deployment type meant for production
environment and uses an external Postgres database for persistence of
data as opposed to H2 for the testing deployment type.

The oss deployment type is a deployment type for users that prefer not
to use the external Google services such as Google Vision and Books.
This version uses instead the public catalogue of the Koninklijke
Bibliotheek and Tesseract for OCR.